Fix SSR
