+++ app/package.json
@@ -59,6 +59,7 @@
     "cookies-next": "^4.3.0",
     "d3-scale-chromatic": "^3.1.0",
     "fast-deep-equal": "^3.1.3",
+    "html-react-parser": "^5.2.2",
     "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
     "maplibre-gl": "^4.7.1",
