+++ app/src/app/components/Cms/RichTextEditor/extensions/Boierplate/BoilerplateContent.tsx
